relocate functions
